New param for new_user
There's a new "send_install_link_via_sms" param in the API.
Add support for that as a fourth (!) positional param, and
also accept a hash reference of named params. In the hash
ref, the new param is called "send_install_link" for brevity.

For consistency's sake, all of the API methods that originally
took positional params will now accept a hashref.

Also: add some helpers around request/response handling. The
main change here is that the last "response" from Authy is stored
in an attribute. This is so that in cases where this module is
obliged to return a bool, but Authy's API returns a complex data
structure, we have a way to get the complex data structure without
changing the module's interface.